Refactor backend to set cookie server side.
Removed original AuthorizationServer and RefreshToken providers, so a
bearer token is no longer passed to the client to be stored in a cookie at
login.

Microsoft Owin security now used to store the claims of the ClaimsIdentity in a cookie.

The previous code used ajaxPrefilter to set the Authentication header. I
could not find a way to remove this header (not unset), and so requests to
Google Maps were failing.